refactor TV app handling for increased resiliency
Until now things could go in not too expected ways *if* when at the time
of homebridge (re)start the TV is on standby.

Now, a given TV will not be fully setup, on (homebridge) first run until
it is ON, so that a full app list can be fetched and then saved to disk.
On subsequent runs if when homebridge restarts the TV is in standby that
cached list will be used. This (hopefully) will help solve the issue of
all the apps becoming sort of orphan and 'unlinked' from TV accessory in
Home app.

Signed-off-by: António Meireles <antonio.meireles@reformi.st>