do not pass errors back to homekit via its callbacks
until now, when we hit an error we just passing it back to homekit via
the callback but when we do that, homekit/home.app just seems to hang
for a while before getting back to normal (afaict this seems to be an
homekit 'feature' and not an homebridge limitation/bug).

from now on, we just print the errors we found to the console and just
not change the state via the callback.

Signed-off-by: António Meireles <antonio.meireles@reformi.st>