properly digest hdmi input changes (closes #20)
if for some reason an user started the plugin for the first time with an
empty HDMI inputs list, and later on added inputs to homebridge's
config.json the plugin would just crash due to some sad logic there now
fixed.

Many thanks to @Marfre888 for helping to identify the core issue and
testing the proper fix.

Signed-off-by: António Meireles <antonio.meireles@reformi.st>