fix plugin initialization so that the TV doesn't turn ON on its own
previously, and in order to fetch the Apps available on the TV we were,
on plugin startup, when the TV was found off briefly turning it on,
which was found to case all sorts of issues namelly when automations are
also in picture.

so, from now on, on plugin startup, TV is ONLY required to be ON the
first time the plugin is run (for a given TV) with the initialization
failing otherwise. subsequent plugin runs will not need the TV to be ON.

should handle #44

Signed-off-by: António Meireles <antonio.meireles@reformi.st>