reset session counter on TVs that have it if a mismatch is found
should solve #67 which was an issue specific to 2018 and later sets that
require crypto for communication.

basically, if a TV is powered down and after powered up the sessionId
is understandably reset, so next time plugin tries to send some command
to the TV everything was going rogue...

  <s:Body>
      <s:Fault>
          <faultcode>s:Client</faultcode>
          <faultstring>UPnPError</faultstring>
          <detail>
              <UPnPError xmlns="urn:schemas-upnp-org:control-1-0">
                <errorCode>612</errorCode>
                <errorDescription>No such session</errorDescription>
                </UPnPError>
            </detail>
    </s:Fault>
  </s:Body>

now when this happens, we just recall requestSessionId() to setup a new
session from scratch and then we retry the failed command.

Signed-off-by: António Meireles <antonio.meireles@reformi.st>