handle #92
when we install this plugin for first time (and nothing exists in
homebridge's config file) homebridge.getPluginConfig() returns undefined
when called by the UI, which wasn't accounted for and broke our custom
UI.

This should now be fixed i.e. an end-user should be able to fully
setup a TV strictly via our UI and without ever having to resort to
manually edit homebridge's config.

Signed-off-by: António Meireles <antonio.meireles@reformi.st>